requirements.txt
python-dotenv
sib-api-v3-sdk==7.6.0
